fro-37-why-zuri-chat-features: removed comments and restyled code base
