+++ packages/main/src/pages/static/whyzurichat/SecondSection.jsx
