+++ packages/main/src/pages/static/whyzurichat/index.js
@@ -1,9 +1,8 @@
 import React from 'react'
-import styles from './why.module.css'
 import Footer from "./footer";
 import {
     TopNavigationBar
-  } from "../../components";
+  } from "../../../components";
 import Body from './Body';
 export default function WhyZuriChat() {
   return (
