+++ packages/main/src/pages/static/whyzurichat/why.module.css
@@ -10,6 +10,7 @@
     margin-right: 1rem;
     
 }
+
 .text {
     width: 100%;
     max-width: 685px;
